Split the remains of OwnerSearch
OwnerSearch::Base and OwnerSearch::Owned are split now too,
leaving only shared functions not to be used from controllers
in Base. Creating a module for it
